pyproject.toml
[tool.poetry]
name = "discord-gateway"
version = "0.1.0"
description = ""
authors = ["Anthony Altieri <anthonyraltieri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.18.0"
aiohttp = "^3.7.4"
"discord.py" = "^1.7.3"
pydantic = "^1.8.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
